# specs/004-core-services-implementation/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: UMOS Core Services API
  version: 1.0.0
  description: >-
    Contract for core service capabilities covering device inventory, telemetry query,
    command tracking, and realtime subscription entry points.
servers:
  - url: http://localhost:8080
    description: Local development
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  parameters:
    DeviceID:
      name: id
      in: path
      required: true
      schema:
        type: integer
      description: Managed device numeric identifier
    DeviceSN:
      name: sn
      in: path
      required: true
      schema:
        type: string
      description: Managed device serial number
    ServiceRequestID:
      name: requestID
      in: path
      required: true
      schema:
        type: string
      description: Tracked service request identifier
  schemas:
    Device:
      type: object
      required: [deviceSN, deviceName, deviceType, vendor, status]
      properties:
        id:
          type: integer
        deviceSN:
          type: string
        deviceName:
          type: string
        deviceType:
          type: string
        vendor:
          type: string
        status:
          type: string
          enum: [unknown, online, offline]
        gatewaySN:
          type: string
          nullable: true
        thingModelID:
          type: integer
          nullable: true
        lastOnlineTime:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateDeviceRequest:
      type: object
      required: [deviceSN, deviceName, deviceType, vendor]
      properties:
        deviceSN:
          type: string
        deviceName:
          type: string
        deviceType:
          type: string
        vendor:
          type: string
        gatewaySN:
          type: string
        thingModelID:
          type: integer
    UpdateDeviceRequest:
      type: object
      properties:
        deviceName:
          type: string
        deviceType:
          type: string
        vendor:
          type: string
        status:
          type: string
          enum: [unknown, online, offline]
        gatewaySN:
          type: string
        thingModelID:
          type: integer
    DeviceListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Device'
        total:
          type: integer
    ServiceRequestCreate:
      type: object
      required: [device_sn, vendor, method]
      properties:
        device_sn:
          type: string
        vendor:
          type: string
        method:
          type: string
        call_type:
          type: string
          enum: [command, property, config]
          default: command
        max_retries:
          type: integer
          minimum: 0
        params:
          type: object
          additionalProperties: true
    ServiceRequest:
      type: object
      required: [id, device_sn, vendor, method, call_type, status, retry_count, max_retries, created_at]
      properties:
        id:
          type: string
        device_sn:
          type: string
        vendor:
          type: string
        method:
          type: string
        call_type:
          type: string
          enum: [command, property, config]
        status:
          type: string
          enum: [pending, sent, success, failed, timeout, retrying]
        tid:
          type: string
        bid:
          type: string
        retry_count:
          type: integer
        max_retries:
          type: integer
        error:
          type: string
        params:
          type: object
          additionalProperties: true
        response:
          type: object
          additionalProperties: true
        sent_at:
          type: string
          format: date-time
          nullable: true
        completed_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
    TelemetryRecord:
      type: object
      required: [deviceSN, messageType, timestamp, payload]
      properties:
        deviceSN:
          type: string
        gatewaySN:
          type: string
          nullable: true
        vendor:
          type: string
        messageType:
          type: string
        measurement:
          type: string
        timestamp:
          type: string
          format: date-time
        payload:
          type: object
          additionalProperties: true
    TelemetryQueryResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/TelemetryRecord'
    RealtimeSubscribeRequest:
      type: object
      required: [topics]
      properties:
        sessionID:
          type: string
        topics:
          type: array
          items:
            type: string
    RealtimeSubscribeAck:
      type: object
      required: [sessionID, topics, accepted]
      properties:
        sessionID:
          type: string
        topics:
          type: array
          items:
            type: string
        accepted:
          type: array
          items:
            type: string
        rejected:
          type: array
          items:
            type: string
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
security:
  - ApiKeyAuth: []
paths:
  /api/v1/devices:
    get:
      summary: List managed devices
      operationId: listDevices
      responses:
        '200':
          description: Managed devices returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceListResponse'
    post:
      summary: Create a managed device
      operationId: createDevice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDeviceRequest'
      responses:
        '201':
          description: Device created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Device'
  /api/v1/devices/{id}:
    get:
      summary: Get device details by numeric ID
      operationId: getDeviceByID
      parameters:
        - $ref: '#/components/parameters/DeviceID'
      responses:
        '200':
          description: Device returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Device'
        '404':
          description: Device not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      summary: Update a managed device by numeric ID
      operationId: updateDeviceByID
      parameters:
        - $ref: '#/components/parameters/DeviceID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateDeviceRequest'
      responses:
        '200':
          description: Device updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Device'
    delete:
      summary: Remove a managed device by numeric ID
      operationId: deleteDeviceByID
      parameters:
        - $ref: '#/components/parameters/DeviceID'
      responses:
        '204':
          description: Device removed
  /api/v1/devices/sn/{sn}:
    get:
      summary: Get device details by serial number
      operationId: getDeviceBySerialNumber
      parameters:
        - $ref: '#/components/parameters/DeviceSN'
      responses:
        '200':
          description: Device returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Device'
        '404':
          description: Device not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/telemetry/{device_sn}:
    get:
      summary: Query historical telemetry for a device
      operationId: queryDeviceTelemetry
      parameters:
        - name: device_sn
          in: path
          required: true
          schema:
            type: string
          description: Managed device serial number
        - name: start
          in: query
          schema:
            type: string
          description: Inclusive start time or relative expression
        - name: end
          in: query
          schema:
            type: string
          description: Inclusive end time or relative expression
      responses:
        '200':
          description: Telemetry returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TelemetryQueryResponse'
  /api/v1/services/call:
    post:
      summary: Submit a device service request
      operationId: createServiceRequest
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServiceRequestCreate'
      responses:
        '202':
          description: Service request accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceRequest'
  /api/v1/services/calls/{id}:
    get:
      summary: Get the current status of a tracked service request
      operationId: getServiceRequest
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Tracked service request identifier
      responses:
        '200':
          description: Service request returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceRequest'
        '404':
          description: Service request not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/services/calls/device/{device_sn}:
    get:
      summary: List service requests for a specific device
      operationId: listServiceRequestsByDevice
      parameters:
        - name: device_sn
          in: path
          required: true
          schema:
            type: string
          description: Device serial number
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
          description: Maximum number of results
      responses:
        '200':
          description: Service requests returned
          content:
            application/json:
              schema:
                type: object
                properties:
                  service_calls:
                    type: array
                    items:
                      $ref: '#/components/schemas/ServiceRequest'
                  total:
                    type: integer
                  page:
                    type: integer
                  page_size:
                    type: integer
  /api/v1/realtime/subscriptions:
    post:
      summary: Register realtime topic subscriptions for the active session
      operationId: subscribeRealtimeTopics
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RealtimeSubscribeRequest'
      responses:
        '200':
          description: Subscription acknowledged
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RealtimeSubscribeAck'
  /health:
    get:
      summary: Basic service health endpoint
      operationId: healthCheck
      responses:
        '200':
          description: Service is alive
  /ready:
    get:
      summary: Operational readiness endpoint
      operationId: readinessCheck
      responses:
        '200':
          description: Service is ready to process traffic
        '503':
          description: Service dependencies are not ready